requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1
django-jazzmin==3.0.1
django-multiselectfield==0.1.13
filelock==3.18.0
platformdirs==4.3.7
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.30.0
